+++ examples/esp_zigbee_touchlink/touchlink_light/main/esp_touchlink_light.c
@@ -34,8 +34,43 @@ void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
     esp_zb_app_signal_type_t sig_type = *p_sg_p;
     esp_zb_zdo_signal_device_annce_params_t *dev_annce_params = NULL;
     switch (sig_type) {
+    case ESP_ZB_ZDO_SIGNAL_SKIP_STARTUP:
+        ESP_LOGI(TAG, "Zigbee stack initialized");
+        esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_INITIALIZATION);
+        break;
+    case ESP_ZB_BDB_SIGNAL_DEVICE_FIRST_START:
+    case ESP_ZB_BDB_SIGNAL_DEVICE_REBOOT:
+        ESP_LOGI(TAG, "Device started up in %s factory-reset mode", esp_zb_bdb_is_factory_new() ? "" : "non");
+        if (esp_zb_bdb_is_factory_new()) {
+            esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_TOUCHLINK_TARGET);
+        } else {
+            ESP_LOGI(TAG, "Device restarted");
+        }
+        break;
     case ESP_ZB_BDB_SIGNAL_TOUCHLINK_TARGET:
-        ESP_LOGI(TAG, "Touchlink target started");
+        ESP_LOGI(TAG, "Touchlink target is ready, awaiting commissioning");
+        break;
+    case ESP_ZB_BDB_SIGNAL_TOUCHLINK_NWK:
+        if (err_status == ESP_OK) {
+            esp_zb_ieee_addr_t extended_pan_id;
+            esp_zb_get_extended_pan_id(extended_pan_id);
+            ESP_LOGI(TAG,
+                     "Commissioning successfully, network information (Extended PAN ID: %02x:%02x:%02x:%02x:%02x:%02x:%02x:%02x, PAN ID: 0x%04hx,"
+                     "Channel:%d, Short Address: 0x%04hx)",
+                     extended_pan_id[7], extended_pan_id[6], extended_pan_id[5], extended_pan_id[4], extended_pan_id[3], extended_pan_id[2],
+                     extended_pan_id[1], extended_pan_id[0], esp_zb_get_pan_id(), esp_zb_get_current_channel(), esp_zb_get_short_address());
+        }
+        break;
+    case ESP_ZB_BDB_SIGNAL_TOUCHLINK_TARGET_FINISHED:
+        if (err_status == ESP_OK) {
+            ESP_LOGI(TAG, "Touchlink target commissioning finished");
+        } else {
+            ESP_LOGI(TAG, "Touchlink target commissioning failed");
+        }
+        break;
+    case ESP_ZB_ZDO_SIGNAL_DEVICE_ANNCE:
+        dev_annce_params = (esp_zb_zdo_signal_device_annce_params_t *)esp_zb_app_signal_get_params(p_sg_p);
+        ESP_LOGI(TAG, "New device commissioned or rejoined (short: 0x%04hx)", dev_annce_params->device_short_addr);
         break;
     case ESP_ZB_NWK_SIGNAL_PERMIT_JOIN_STATUS:
         if (err_status == ESP_OK) {
@@ -45,20 +80,6 @@ void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
                 ESP_LOGW(TAG, "Network(0x%04hx) closed, devices joining not allowed.", esp_zb_get_pan_id());
             }
         }
-        break;
-    case ESP_ZB_BDB_SIGNAL_TOUCHLINK_NWK:
-        ESP_LOGI(TAG, "Touchlink target : network started");
-        break;
-    case ESP_ZB_BDB_SIGNAL_TOUCHLINK_TARGET_FINISHED:
-        if (err_status == ESP_OK) {
-            ESP_LOGI(TAG, "Touchlink target finished: success");
-        } else {
-            ESP_LOGI(TAG, "Touchlink target finished: failed");
-        }
-        break;
-    case ESP_ZB_ZDO_SIGNAL_DEVICE_ANNCE:
-        dev_annce_params = (esp_zb_zdo_signal_device_annce_params_t *)esp_zb_app_signal_get_params(p_sg_p);
-        ESP_LOGI(TAG, "New device commissioned or rejoined (short: 0x%04hx)", dev_annce_params->device_short_addr);
         break;
     default:
         ESP_LOGI(TAG, "ZDO signal: %s (0x%x), status: %s", esp_zb_zdo_signal_to_string(sig_type), sig_type, esp_err_to_name(err_status));
@@ -107,11 +128,11 @@ static void esp_zb_task(void *pvParameters)
     /* Initialize Zigbee stack with Zigbee coordinator config */
     esp_zb_cfg_t zb_nwk_cfg = ESP_ZB_ZR_CONFIG();
     esp_zb_init(&zb_nwk_cfg);
+
     esp_zb_set_channel_mask(ESP_ZB_TOUCHLINK_CHANNEL_MASK);
     esp_zb_set_rx_on_when_idle(true);
- 
-    esp_zb_set_bdb_commissioning_mode(ESP_ZB_BDB_TOUCHLINK_TARGET);
-    
+    esp_zb_zdo_touchlink_target_set_timeout(TOUCHLINK_TARGET_TIMEOUT);
+
     esp_zb_attribute_list_t *touchlink_cluster = esp_zb_touchlink_commissioning_cluster_create();
 
     esp_zb_on_off_light_cfg_t light_cfg = ESP_ZB_DEFAULT_ON_OFF_LIGHT_CONFIG();
@@ -128,7 +149,7 @@ static void esp_zb_task(void *pvParameters)
     esp_zb_device_register(esp_zb_ep_list);
     esp_zb_core_action_handler_register(zb_action_handler);
 
-    ESP_ERROR_CHECK(esp_zb_start(true));
+    ESP_ERROR_CHECK(esp_zb_start(false));
     esp_zb_main_loop_iteration();
 }